printing_wizard/pyproject.toml
[project]
name = "label-printer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "freetype-py>=2.5.1",
    "matplotlib>=3.10.9",
    "shapely>=2.1.2",
    "svgpathtools>=1.7.2",
    "pyserial"
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
